Re-factor SDP parser into three sepearte files
